+++ example/lib/main.dart
@@ -10,7 +10,7 @@ void main() => runApp(MyApp());
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final bool isIos = Theme.of(context).platform == TargetPlatform.iOS;
+    final isIos = Theme.of(context).platform == TargetPlatform.iOS;
 
     const Widget aboutPage = AboutPage(
       title: Text('About'),